fix: fix the emoji message size once again
this used to work once upon a time, but it stopped working recently.
This commit helps to make the emoji-only-messages bigger. In case, if emoji's
are sent as a reply to a text message, this does not affect size then.

fix: fix the size of the `(edited)` string when the message
contains only emojis' and also when its edited