Don't show option to private chat with oneself
While this option works, it doesn't make much sense since there isn't
much of a good reason to talk to oneself aside from for testing
purposes. The created conversation also has confusing UI, showing as an
"empty room", not a DM. If needed, creating an empty room with E2E
encryption has the same effect as creating a private chat with
yourself.

Closes #446.