Add Windows-specific screenshots as custom values
This is clearly a hack. It would be preferable to add those directly
to the existing list of screenshots with some property marking them as
screenshots of the Windows version for the Microsoft Store
(https://github.com/ximion/appstream/issues/390). The IDs are derived
from the IDs used by the CSV export of an app submission in Microsoft
Partner Center.