Add features as custom values
Microsoft Store allows adding a list of up-to 20 features to an app
submission. It's not clear whether this adds value over the common
practice for KDE apps to add a list of features to the description.
If the search in Microsoft Store would rank matching features higher
than a full-text match in the description, then it might make sense.

The problem is that this would duplicate the list of features in
Microsoft Store because removing the list of features from the description
would remove the features from the app listing in Linux app stores.
Therefore, we should probably stick to the established practice and leave
the list of features in Microsoft Store empty.

See also https://github.com/ximion/appstream/issues/388