Ensure that room list does not show rooms without title during startup
There's a brief moment during startup where the model knows about the rooms, but their state
is not loaded, which makes them show up in the room list with ugly fallback titles.
To prevent this, we delay showing the treeview until the rooms' base state is loaded.

Also add a loading placeholder and hide the default placeholder during this phase.